+++ kasper-common/src/main/java/com/viadeo/kasper/tools/ObjectMapperProvider.java
@@ -77,7 +77,7 @@ public final class ObjectMapperProvider {
 
         mapper.registerModule(kasperClientModule);
         mapper.registerModule(new KasperImmutabilityParanamerModule());
-        mapper.registerModule(new LowerEnumModule());
+        
 
         /* Third-party modules */
         mapper.registerModule(new GuavaModule());