+++ kasper-documentation/src/main/java/com/viadeo/kasper/doc/element/DocumentedDomain.java
@@ -15,6 +15,7 @@ import com.viadeo.kasper.cqrs.query.CollectionQueryResult;
 import com.viadeo.kasper.cqrs.query.QueryResult;
 import com.viadeo.kasper.doc.initializer.DocumentedElementVisitor;
 import com.viadeo.kasper.doc.nodes.DocumentedBean;
+import com.viadeo.kasper.event.IEvent;
 import com.viadeo.kasper.tools.ReflectionGenericsResolver;
 
 import java.lang.reflect.Field;
@@ -88,6 +89,11 @@ public class DocumentedDomain extends AbstractElement {
 
         final Map<Class, DocumentedEvent> events = Maps.newHashMap();
         final Map<Class, DocumentedConcept> concepts = Maps.newHashMap();
+
+        for (final Class<? extends IEvent> eventClass : domainDescriptor.getEventClasses()) {
+            DocumentedEvent documentedEvent = new DocumentedEvent(this, null, eventClass);
+            events.put(eventClass, documentedEvent);
+        }
 
         for (final RepositoryDescriptor descriptor : domainDescriptor.getRepositoryDescriptors()) {
             final DocumentedRepository documentedRepository = new DocumentedRepository(this, descriptor);