+++ kasper-platform/src/main/java/com/viadeo/kasper/client/platform/domain/descriptor/DomainDescriptorFactory.java
@@ -21,10 +21,17 @@ import com.viadeo.kasper.ddd.repository.Repository;
 import com.viadeo.kasper.er.Relation;
 import com.viadeo.kasper.event.Event;
 import com.viadeo.kasper.event.EventListener;
+import com.viadeo.kasper.event.IEvent;
 import com.viadeo.kasper.tools.ReflectionGenericsResolver;
 import org.axonframework.eventhandling.annotation.EventHandler;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.context.annotation.ClassPathScanningCandidateComponentProvider;
+import org.springframework.core.type.filter.AssignableTypeFilter;
 
 import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -32,6 +39,8 @@ import java.util.Map;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 public class DomainDescriptorFactory {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(DomainDescriptorFactory.class);
 
     private static final Function<CommandHandler, CommandHandlerDescriptor> TO_COMMAND_HANDLER_DESCRIPTOR_FUNCTION = new Function<CommandHandler, CommandHandlerDescriptor>() {
         @Override
@@ -70,7 +79,8 @@ public class DomainDescriptorFactory {
             Collections2.transform(domainBundle.getQueryHandlers(), TO_QUERY_HANDLER_DESCRIPTOR_FUNCTION),
             Collections2.transform(domainBundle.getCommandHandlers(), TO_COMMAND_HANDLER_DESCRIPTOR_FUNCTION),
             Collections2.transform(domainBundle.getRepositories(), TO_REPOSITORY_DESCRIPTOR_FUNCTION),
-            Collections2.transform(domainBundle.getEventListeners(), TO_EVENT_LISTENER_DESCRIPTOR_FUNCTION)
+            Collections2.transform(domainBundle.getEventListeners(), TO_EVENT_LISTENER_DESCRIPTOR_FUNCTION),
+            retrieveEventsFrom(domainBundle.getDomain().getClass())
         );
     }
 
@@ -86,11 +96,33 @@ public class DomainDescriptorFactory {
             Collections2.transform(checkNotNull(queryHandlers), TO_QUERY_HANDLER_DESCRIPTOR_FUNCTION),
             Collections2.transform(checkNotNull(commandHandlers), TO_COMMAND_HANDLER_DESCRIPTOR_FUNCTION),
             Collections2.transform(checkNotNull(repositories), TO_REPOSITORY_DESCRIPTOR_FUNCTION),
-            Collections2.transform(checkNotNull(eventListeners), TO_EVENT_LISTENER_DESCRIPTOR_FUNCTION)
+            Collections2.transform(checkNotNull(eventListeners), TO_EVENT_LISTENER_DESCRIPTOR_FUNCTION),
+            retrieveEventsFrom(domainClass)
         );
     }
 
     // ------------------------------------------------------------------------
+
+    public static Collection<Class<? extends IEvent>> retrieveEventsFrom(final Class<? extends Domain> domainClass) {
+        final List<Class<? extends IEvent>> eventClasses = Lists.newArrayList();
+
+        final ClassPathScanningCandidateComponentProvider scanner = new ClassPathScanningCandidateComponentProvider(false);
+        scanner.addIncludeFilter(new AssignableTypeFilter(IEvent.class));
+
+        for (final BeanDefinition bd : scanner.findCandidateComponents(domainClass.getPackage().getName())) {
+            try {
+                @SuppressWarnings("unchecked")
+                Class<IEvent> eventClass = (Class<IEvent>) Class.forName(bd.getBeanClassName());
+                if ( ! (Modifier.isAbstract(eventClass.getModifiers()) || Modifier.isInterface(eventClass.getModifiers())) ) {
+                    eventClasses.add(eventClass);
+                }
+            } catch (ClassNotFoundException e) {
+                LOGGER.error("Failed to found class : '{}'", bd.getBeanClassName());
+            }
+        }
+
+        return eventClasses;
+    }
 
     @SuppressWarnings("unchecked")
     public static CommandHandlerDescriptor toCommandHandlerDescriptor(final CommandHandler commandHandler) {