+++ kasper-test/src/main/java/com/viadeo/kasper/test/platform/KasperResultValidator.java
@@ -11,8 +11,6 @@ import org.axonframework.test.ResultValidator;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.Matcher;
 
-import java.util.List;
-
 import static com.google.common.base.Preconditions.checkNotNull;
 
 public class KasperResultValidator implements ResultValidator {
@@ -29,7 +27,7 @@ public class KasperResultValidator implements ResultValidator {
     }
 
     @Override
-    public ResultValidator expectEventsMatching(final Matcher<List<?>> matcher) {
+    public ResultValidator expectEventsMatching(final Matcher<? extends Iterable<?>> matcher) {
         return delegate.expectEventsMatching(matcher);
     }
 
@@ -39,7 +37,7 @@ public class KasperResultValidator implements ResultValidator {
     }
 
     @Override
-    public ResultValidator expectPublishedEventsMatching(final Matcher<List<?>> matcher) {
+    public ResultValidator expectPublishedEventsMatching(final Matcher<? extends Iterable<?>> matcher) {
         return delegate.expectPublishedEventsMatching(matcher);
     }
 
@@ -49,7 +47,7 @@ public class KasperResultValidator implements ResultValidator {
     }
 
     @Override
-    public ResultValidator expectStoredEventsMatching(final Matcher<List<?>> matcher) {
+    public ResultValidator expectStoredEventsMatching(final Matcher<? extends Iterable<?>> matcher) {
         return delegate.expectStoredEventsMatching(matcher);
     }